pyproject.toml
[tool.poetry]
name = "therockbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
Flask = "^2.0.2"
replit = "^3.2.4"
buttons = "^0.1.9"
discord-buttons-plugin = "^0.1.2"
Command = "^0.1.0"
discord-py-interactions = "^3.0.2"
orjson = "^3.6.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
